cambios

+++ app/Http/Controllers/categoriaClienteController.php
@@ -64,16 +64,17 @@ class categoriaClienteController extends Controller
             if (isset(Auth::user()->empresa->grupo)) {
                 foreach (Auth::user()->empresa->grupo->usuarios->empresas as $empresas) {
                     if (Auth::user()->empresa->empresa_id != $empresas->empresa->empresa_id) {
-                        $categoriaClien = new Categoria_Cliente();
-                        $categoriaClien->categoria_cliente_nombre = $request->get('categoria_cliente_nombre');
-                        $categoriaClien->categoria_cliente_descripcion = $request->get('categoria_cliente_descripcion');
-                        $categoriaClien->empresa_id =$empresas->empresa->empresa_id;
-                        $categoriaClien->categoria_cliente_estado = 1;
-                        $categoriaClien->save();
-                        /*Inicio de registro de auditoria */
-                        $auditoria = new generalController();
-                        $auditoria->registrarAuditoria('Registro de categoria cliente -> '.$request->get('categoria_cliente_nombre').' con descripcion -> '.$request->get('categoria_cliente_descripcion').' Con empresa ruc '.$empresas->empresa->empresa_ruc.' Con razon social'.$empresas->empresa->empresa_razonSocial,'0','');
-                      
+                        if (Auth::user()->empresa->grupo->grupo_duplicado == '1') {
+                            $categoriaClien = new Categoria_Cliente();
+                            $categoriaClien->categoria_cliente_nombre = $request->get('categoria_cliente_nombre');
+                            $categoriaClien->categoria_cliente_descripcion = $request->get('categoria_cliente_descripcion');
+                            $categoriaClien->empresa_id =$empresas->empresa->empresa_id;
+                            $categoriaClien->categoria_cliente_estado = 1;
+                            $categoriaClien->save();
+                            /*Inicio de registro de auditoria */
+                            $auditoria = new generalController();
+                            $auditoria->registrarAuditoria('Registro de categoria cliente -> '.$request->get('categoria_cliente_nombre').' con descripcion -> '.$request->get('categoria_cliente_descripcion').' Con empresa ruc '.$empresas->empresa->empresa_ruc.' Con razon social'.$empresas->empresa->empresa_razonSocial,'0','');
+                        }
                        
                     }
                 }
@@ -162,17 +163,19 @@ class categoriaClienteController extends Controller
             if (isset(Auth::user()->empresa->grupo)) {
                 foreach (Auth::user()->empresa->grupo->usuarios->empresas as $empresas) {
                     if (Auth::user()->empresa->empresa_id != $empresas->empresa->empresa_id) {
-                        $categoriaClienaux = Categoria_Cliente::CategoriaClienteEmpresaNombre($categoriaClienaux->categoria_cliente_nombre,$empresas->empresa->empresa_id)->first();
-                        $categoriaClien = Categoria_Cliente::findOrFail($categoriaClienaux->categoria_cliente_id);
-                        $categoriaClien->categoria_cliente_nombre = $request->get('categoria_cliente_nombre');
-                        $categoriaClien->categoria_cliente_descripcion = $request->get('categoria_cliente_descripcion');          
-                        if ($request->get('categoria_cliente_estado') == "on"){
-                            $categoriaClien->categoria_cliente_estado = 1;
-                        }else{
-                            $categoriaClien->categoria_cliente_estado = 0;
+                        if (Auth::user()->empresa->grupo->grupo_duplicado == '1') {
+                            $categoriaClienaux = Categoria_Cliente::CategoriaClienteEmpresaNombre($categoriaClienaux->categoria_cliente_nombre,$empresas->empresa->empresa_id)->first();
+                            $categoriaClien = Categoria_Cliente::findOrFail($categoriaClienaux->categoria_cliente_id);
+                            $categoriaClien->categoria_cliente_nombre = $request->get('categoria_cliente_nombre');
+                            $categoriaClien->categoria_cliente_descripcion = $request->get('categoria_cliente_descripcion');          
+                            if ($request->get('categoria_cliente_estado') == "on"){
+                                $categoriaClien->categoria_cliente_estado = 1;
+                            }else{
+                                $categoriaClien->categoria_cliente_estado = 0;
+                            }
+                            $categoriaClien->save();
+                            /*Inicio de registro de auditoria */
                         }
-                        $categoriaClien->save();
-                        /*Inicio de registro de auditoria */
                         
                     }
                 }
@@ -206,14 +209,15 @@ class categoriaClienteController extends Controller
             if (isset(Auth::user()->empresa->grupo)) {
                 foreach (Auth::user()->empresa->grupo->usuarios->empresas as $empresas) {
                     if (Auth::user()->empresa->empresa_id != $empresas->empresa->empresa_id) {
-                        $categoriaClienaux = Categoria_Cliente::CategoriaClienteEmpresaNombre($categoriaClienaux->categoria_cliente_nombre,$empresas->empresa->empresa_id)->first();
-                        $categoriaClien = Categoria_Cliente::findOrFail($categoriaClienaux->categoria_cliente_id);
-                        $categoriaClien->delete();
-                        /*Inicio de registro de auditoria */
-                        $auditoria = new generalController();
-                        $auditoria->registrarAuditoria('Eliminacion de categoria cliente -> '.$categoriaClien->categoria_cliente_nombre.' con descripcion -> '.$categoriaClien->categoria_cliente_descripcion.' Con empresa ruc '.$empresas->empresa->empresa_ruc.' Con razon social'.$empresas->empresa->empresa_razonSocial,'0','');
-                        /*Inicio de registro de auditoria */
-                        
+                        if (Auth::user()->empresa->grupo->grupo_duplicado == '1') {
+                            $categoriaClienaux = Categoria_Cliente::CategoriaClienteEmpresaNombre($categoriaClienaux->categoria_cliente_nombre,$empresas->empresa->empresa_id)->first();
+                            $categoriaClien = Categoria_Cliente::findOrFail($categoriaClienaux->categoria_cliente_id);
+                            $categoriaClien->delete();
+                            /*Inicio de registro de auditoria */
+                            $auditoria = new generalController();
+                            $auditoria->registrarAuditoria('Eliminacion de categoria cliente -> '.$categoriaClien->categoria_cliente_nombre.' con descripcion -> '.$categoriaClien->categoria_cliente_descripcion.' Con empresa ruc '.$empresas->empresa->empresa_ruc.' Con razon social'.$empresas->empresa->empresa_razonSocial,'0','');
+                            /*Inicio de registro de auditoria */
+                        }
                     }
                 }
             }
